+++ Hypercube.Shared.Math/Transform/Transform2.cs
@@ -56,7 +56,7 @@ public partial struct Transform2 : ITransform
     private void UpdateMatrix()
     {
         Matrix = Matrix4X4.CreateTranslation(Position) *
-                 Matrix4X4.CreateRotationZ(Rotation) *
+                 Matrix4X4.CreateRotationZ((float)Rotation) *
                  Matrix4X4.CreateScale(Scale);
     }
 }
